+++ src/shared/icons/index.ts
@@ -1,5 +1,6 @@
 export { default as NebulaIcon } from './NebulaIcon'
 export { default as MissionIcon } from './MissionIcon'
+export { default as CampfireIcon } from './CampfireIcon'
 export { default as NasaIcon } from './NasaIcon'
 export { default as AuroraIcon } from './AuroraIcon'
 export { default as LunaIcon } from './LunaIcon'